[libyafuse2]: fix core dump when ext4 is mounted due to incorrect type of 'size_t'.

+++ libyafuse2/include/libfat/libfat.h
@@ -97,7 +97,7 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
 int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int32_t *dentries);
 int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32_t dentries, struct msdos_dir_slot *dslot, struct msdos_dir_entry *dentry);
 int32_t fat_dent_attr_is_dir(const struct msdos_dir_entry *dentry, int32_t *status);
-int32_t fat_fill_file(const struct fat_super_block *sb, int32_t cluster, size_t size, uint8_t *buf);
+int32_t fat_fill_file(const struct fat_super_block *sb, int32_t cluster, int64_t size, uint8_t *buf);
 
 void fat_show_stats(const struct fat_super_block *sb);
 void fat_show_dslot(const struct fat_super_block *sb, const struct msdos_dir_slot *dslot);
