+++ libyafuse2/include/libio/io.h
@@ -45,8 +45,8 @@
  */
 int32_t io_open(const char *fs_name);
 void io_close(void);
-int32_t io_seek(long offset);
-int32_t io_read(uint8_t *data, size_t len);
-int32_t io_write(uint8_t *data, size_t len);
+int32_t io_seek(int64_t offset);
+int32_t io_read(uint8_t *data, int64_t len);
+int32_t io_write(uint8_t *data, int64_t len);
 
 #endif /* _IO_H */
