+++ libyafuse2/libfat/dir.c
@@ -98,9 +98,9 @@ int32_t fat_fill_dent_start(const struct fat_super_block *sb, const struct msdos
 int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentries)
 {
   int32_t root_den_sec = 0;
-  int32_t offset = 0;
+  int64_t offset = 0;
   struct msdos_dir_entry dentry;
-  int32_t sz = 0;
+  int64_t sz = 0;
   int32_t i = 0;
   int32_t ret = 0;
 
@@ -111,14 +111,14 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
 
   offset = root_den_sec * (int32_t)GET_UNALIGNED_LE16(sb->bs.sector_size);
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
 
   sz = sizeof(struct msdos_dir_entry);
 
-  ret = io_read((uint8_t *)&dentry, (size_t)sz);
+  ret = io_read((uint8_t *)&dentry, sz);
   if (ret != 0) {
     return -1;
   }
@@ -127,7 +127,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
    * Long File Names (LFN), i.e., dslot, is used
    */
   if (dentry.attr == ATTR_EXT) {
-    ret = io_read((uint8_t *)&dentry, (size_t)sz);
+    ret = io_read((uint8_t *)&dentry, sz);
     if (ret != 0) {
       return -1;
     }
@@ -136,7 +136,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
   while (dentry.name[0] != '\0') {
     ++i;
 
-    ret = io_read((uint8_t *)&dentry, (size_t)sz);
+    ret = io_read((uint8_t *)&dentry, sz);
     if (ret != 0) {
       return -1;
     }
@@ -145,7 +145,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dentry.name[0] != '\0' && dentry.attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry, (size_t)sz);
+      ret = io_read((uint8_t *)&dentry, sz);
       if (ret != 0) {
         return -1;
       }
@@ -160,8 +160,8 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
 int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries, struct msdos_dir_slot *dslot, struct msdos_dir_entry *dentry)
 {
   int32_t root_den_sec = 0;
-  int32_t offset = 0;
-  int32_t sz = 0;
+  int64_t offset = 0;
+  int64_t sz = 0;
   int32_t i = 0;
   int32_t ret = 0;
 
@@ -172,7 +172,7 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
 
   offset = root_den_sec * (int32_t)GET_UNALIGNED_LE16(sb->bs.sector_size);
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
@@ -180,7 +180,7 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
   sz = sizeof(struct msdos_dir_entry);
 
   for (i = 0; i < dentries; ++i) {
-    ret = io_read((uint8_t *)&dslot[i], (size_t)sz);
+    ret = io_read((uint8_t *)&dslot[i], sz);
     if (ret != 0) {
       break;
     }
@@ -189,7 +189,7 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dslot[i].attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry[i], (size_t)sz);
+      ret = io_read((uint8_t *)&dentry[i], sz);
       if (ret != 0) {
         break;
       }
@@ -204,9 +204,9 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
 int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int32_t *dentries)
 {
   int32_t sector = 0;
-  int32_t offset = 0;
+  int64_t offset = 0;
   struct msdos_dir_entry dentry;
-  int32_t sz = 0;
+  int64_t sz = 0;
   int32_t i = 1;
   int32_t ret = 0;
 
@@ -217,7 +217,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
 
   offset = sector * (int32_t)GET_UNALIGNED_LE16(sb->bs.sector_size);
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
@@ -227,7 +227,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
   /*
    * Read FAT dentry of '.'
    */
-  ret = io_read((uint8_t *)&dentry, (size_t)sz);
+  ret = io_read((uint8_t *)&dentry, sz);
   if (ret != 0) {
     return -1;
   }
@@ -235,7 +235,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
   /*
    * Read FAT dentry of '..'
    */
-  ret = io_read((uint8_t *)&dentry, (size_t)sz);
+  ret = io_read((uint8_t *)&dentry, sz);
   if (ret != 0) {
     return -1;
   }
@@ -243,7 +243,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
   while (dentry.name[0] != '\0') {
     ++i;
 
-    ret = io_read((uint8_t *)&dentry, (size_t)sz);
+    ret = io_read((uint8_t *)&dentry, sz);
     if (ret != 0) {
       return -1;
     }
@@ -252,7 +252,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dentry.attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry, (size_t)sz);
+      ret = io_read((uint8_t *)&dentry, sz);
       if (ret != 0) {
         return -1;
       }
@@ -267,8 +267,8 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
 int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32_t dentries, struct msdos_dir_slot *dslot, struct msdos_dir_entry *dentry)
 {
   int32_t sector = 0;
-  int32_t offset = 0;
-  int32_t sz = 0;
+  int64_t offset = 0;
+  int64_t sz = 0;
   int32_t i = 0;
   int32_t ret = 0;
 
@@ -279,7 +279,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
 
   offset = sector * (int32_t)GET_UNALIGNED_LE16(sb->bs.sector_size);
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
@@ -291,7 +291,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
    */
   memset((void *)&dslot[0], 0, sz);
 
-  ret = io_read((uint8_t *)&dentry[0], (size_t)sz);
+  ret = io_read((uint8_t *)&dentry[0], sz);
   if (ret != 0) {
     return -1;
   }
@@ -301,7 +301,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
    */
   memset((void *)&dslot[1], 0, sz);
 
-  ret = io_read((uint8_t *)&dentry[1], (size_t)sz);
+  ret = io_read((uint8_t *)&dentry[1], sz);
   if (ret != 0) {
     return -1;
   }
@@ -310,7 +310,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
    * Fill in other dentry
    */
   for (i = 2; i < dentries; ++i) {
-    ret = io_read((uint8_t *)&dslot[i], (size_t)sz);
+    ret = io_read((uint8_t *)&dslot[i], sz);
     if (ret != 0) {
       break;
     }
@@ -319,7 +319,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dslot[i].attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry[i], (size_t)sz);
+      ret = io_read((uint8_t *)&dentry[i], sz);
       if (ret != 0) {
         break;
       }
