+++ libyafuse2/libfat/file.c
@@ -54,10 +54,10 @@
 /*
  * Function Definition
  */
-int32_t fat_fill_file(const struct fat_super_block *sb, int32_t cluster, size_t size, uint8_t *buf)
+int32_t fat_fill_file(const struct fat_super_block *sb, int32_t cluster, int64_t size, uint8_t *buf)
 {
   int32_t sector = 0;
-  int32_t offset = 0;
+  int64_t offset = 0;
   int32_t ret = 0;
 
   ret = fat_fill_clus2sec(sb, cluster, &sector);
@@ -67,12 +67,12 @@ int32_t fat_fill_file(const struct fat_super_block *sb, int32_t cluster, size_t 
 
   offset = sector * (int32_t)GET_UNALIGNED_LE16(sb->bs.sector_size);
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
 
-  ret = io_read(buf, (size_t)size);
+  ret = io_read(buf, size);
   if (ret != 0) {
     return -1;
   }
