+++ libyafuse2/libfat/super.c
@@ -78,8 +78,8 @@ static int32_t fat_is_valid_media(uint32_t media)
 
 int32_t fat_fill_sb(struct fat_super_block *sb)
 {
-  int32_t offset = 0;
-  size_t sz = 0;
+  int64_t offset = 0;
+  int64_t sz = 0;
   int32_t is_fat32_fs = 0;
   int32_t ret = 0;
 
@@ -87,13 +87,13 @@ int32_t fat_fill_sb(struct fat_super_block *sb)
    * Fill in FAT boot sector
    */
   offset = 0;
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
 
   sz = sizeof(struct fat_boot_sector);
-  ret = io_read((uint8_t *)&sb->bs, (size_t)sz);
+  ret = io_read((uint8_t *)&sb->bs, sz);
   if (ret != 0) {
     memset((void *)&sb->bs, 0, sz);
     return -1;
@@ -121,13 +121,13 @@ int32_t fat_fill_sb(struct fat_super_block *sb)
     offset = FAT16_BSX_OFFSET;
   }
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
 
   sz = sizeof(struct fat_boot_bsx);
-  ret = io_read((uint8_t *)&sb->bb, (size_t)sz);
+  ret = io_read((uint8_t *)&sb->bb, sz);
   if (ret != 0) {
     memset((void *)&sb->bb, 0, sz);
     return -1;
@@ -139,13 +139,13 @@ int32_t fat_fill_sb(struct fat_super_block *sb)
      */
     offset = sb->bs.info_sector == 0 ? GET_UNALIGNED_LE16(sb->bs.sector_size) : sb->bs.info_sector * GET_UNALIGNED_LE16(sb->bs.sector_size);
 
-    ret = io_seek((long)offset);
+    ret = io_seek(offset);
     if (ret != 0) {
       return -1;
     }
 
     sz = sizeof(struct fat_boot_fsinfo);
-    ret = io_read((uint8_t *)&sb->bf, (size_t)sz);
+    ret = io_read((uint8_t *)&sb->bf, sz);
     if (ret != 0) {
       memset((void *)&sb->bf, 0, sz);
       return -1;
